test(svelte-query/createQuery): move test files into 'createQuery' directory
